+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## Changelog
 
 ## v1.0.0
-- Initial version, includes FormBuilder mixin, FormItem class, Validator class and ValidatorResult class
+- Initial release